Ghosts and Pacman are central on field

+++ src/Ghost.hpp
@@ -16,7 +16,7 @@ class Ghost : public GameCharacter
 {
 public:
     
-    Ghost(int planeWidth, int planeDepth, int posX, int posZ, PhongShader* pPhongShader);
+    Ghost(int planeWidth, int planeDepth, float posX, float posZ, PhongShader* pPhongShader);
     
     /**
      STEERING
